requirements.txt
numpy==1.21.6
pandas==1.3.5
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
six==1.17.0
tabulate==0.9.0
